pyproject.toml
[project]
name = "learning-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-generativeai>=0.8.6",
    "langchain>=1.2.7",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.7",
    "langchain-google-genai>=4.2.0",
    "langchain-huggingface>=1.2.0",
    "langgraph>=1.0.7",
    "pymongo>=4.16.0",
    "pypdf>=6.6.2",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "sentence-transformers>=5.2.2",
    "uvicorn>=0.40.0",
]
